Se corrigió la operación de borrar

+++ src/main/java/dev/j3c/restapi/servicio/ServiciosPersona.java
@@ -30,7 +30,10 @@ public class ServiciosPersona implements InterfazServiciosPersona{
 
     @Override
     public void borrar(int id) {
-        data.deleteById(id);
+        boolean existe = data.existsById(id);
+        if(existe) {
+            data.deleteById(id);
+        }
     }
 
     @Override
